Rev 146

+++ acuity-api/src/main/java/com/acuity/rs/api/RSBuffer.java
@@ -9,45 +9,45 @@ public interface RSBuffer extends RSNode {
 
 	byte[] getPayload();
 
-	void invokeDecryptXtea(int[] var0, int var1);
+	void invokeDecryptXtea(int[] var0, byte var1);
 
-	void invokeEncryptXtea(int[] var0, int var1, int var2, byte var3);
+	void invokeEncryptXtea(int[] var0, int var1, int var2, int var3);
 
 	void invokeEncryptXtea2(int[] var0, int var1);
 
-	void invokePut24bitInt(int var0, byte var1);
+	void invokePut24bitInt(int var0, int var1);
 
-	void invokePutByte(int var0, int var1);
+	void invokePutByte(int var0, byte var1);
 
-	void invokePutBytes(byte[] var0, int var1, int var2, byte var3);
+	void invokePutBytes(byte[] var0, int var1, int var2, int var3);
 
-	void invokePutInt(int var0, int var1);
+	void invokePutInt(int var0, byte var1);
 
 	void invokePutLong(long var0);
 
-	void invokePutShort(int var0, byte var1);
+	void invokePutShort(int var0, int var1);
 
 	void invokePutVarInt(int var0, int var1);
 
-	int invokeRead24BitInt(byte var0);
+	int invokeRead24BitInt(short var0);
 
-	byte invokeReadByte(byte var0);
+	byte invokeReadByte(int var0);
 
 	void invokeReadBytes(byte[] var0, int var1, int var2, int var3);
 
-	int invokeReadInt(byte var0);
+	int invokeReadInt(int var0);
 
 	long invokeReadLong(int var0);
 
 	int invokeReadShort(int var0);
 
-	int invokeReadShortSmart(int var0);
+	int invokeReadShortSmart(byte var0);
 
-	java.lang.String invokeReadString(int var0);
+	java.lang.String invokeReadString(byte var0);
 
-	int invokeReadUnsignedByte(byte var0);
+	int invokeReadUnsignedByte(int var0);
 
-	int invokeReadUnsignedShort(int var0);
+	int invokeReadUnsignedShort(short var0);
 
 	int invokeReadVarInt(int var0);
