+++ acuity-api/src/main/java/com/acuity/rs/api/RSKitDefinition.java
@@ -19,7 +19,7 @@ public interface RSKitDefinition extends RSCacheableNode {
 
 	short[] getRetextureToReplace();
 
-	com.acuity.rs.api.RSModelComposite invokeGetModelData(byte var0);
+	com.acuity.rs.api.RSModelComposite invokeGetModelData(int var0);
 
 	boolean invokeReady(int var0);
