+++ acuity-api/src/main/java/com/acuity/rs/api/RSMainBufferProvider.java
@@ -7,9 +7,9 @@ public interface RSMainBufferProvider extends RSBufferProvider {
 
 	java.awt.Image getImage();
 
-	void invokeDraw(java.awt.Graphics var0, int var1, int var2, byte var3);
+	void invokeDraw(java.awt.Graphics var0, int var1, int var2, short var3);
 
-	void invokeDrawSub(java.awt.Graphics var0, int var1, int var2, int var3, int var4, short var5);
+	void invokeDrawSub(java.awt.Graphics var0, int var1, int var2, int var3, int var4, int var5);
 
 	void setImage(java.awt.Image var0);
 }
