+++ acuity-api/src/main/java/com/acuity/rs/api/RSPacketBuffer.java
@@ -9,11 +9,11 @@ public interface RSPacketBuffer extends RSBuffer {
 
 	com.acuity.rs.api.RSISAACCipher getCipher();
 
-	int invokeReadOpcode(byte var0);
+	int invokeReadOpcode(int var0);
 
 	void invokeSeed(int[] var0, int var1);
 
-	void invokeWriteHeader(int var0, byte var1);
+	void invokeWriteHeader(int var0, int var1);
 
 	void setBitOffset(int var0);
