+++ acuity-api/src/main/java/com/acuity/rs/api/RSRenderOverview.java
@@ -7,5 +7,5 @@ public interface RSRenderOverview {
 
 	void invokeExtractData(int var0, int var1, int var2, int var3, byte var4);
 
-	void invokeExtractWorldmap(int var0, int var1, int var2, int var3, int var4, int var5);
+	void invokeExtractWorldmap(int var0, int var1, int var2, int var3, int var4, byte var5);
 }
